# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Transferências e Usuários",
    "version": "1.0.0",
    "description": "API para login, registro, consulta de usuários e transferências."
  },
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "favorecidos": { "type": "array", "items": { "type": "string" } }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário registrado com sucesso." },
          "400": { "description": "Usuário e senha são obrigatórios." },
          "409": { "description": "Usuário já existe." }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login realizado com sucesso." },
          "400": { "description": "Usuário e senha são obrigatórios." },
          "401": { "description": "Usuário ou senha inválidos." }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Listar usuários",
        "responses": {
          "200": { "description": "Lista de usuários." }
        }
      }
    },
    "/transfers": {
      "post": {
        "summary": "Realizar transferência",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "remetente": { "type": "string" },
                  "destinatario": { "type": "string" },
                  "valor": { "type": "number" }
                },
                "required": ["remetente", "destinatario", "valor"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Transferência realizada com sucesso." },
          "400": { "description": "Remetente, destinatário e valor são obrigatórios." },
          "403": { "description": "Transferências para não favorecidos só podem ser menores que R$ 5.000,00." },
          "404": { "description": "Remetente ou destinatário não encontrado." }
        }
      },
      "get": {
        "summary": "Listar transferências",
        "responses": {
          "200": { "description": "Lista de transferências." }
        }
      }
    }
  }
}